fix filenames + lint errors
apparently a change in casing isnt tracked accross git if it also doesnt matter to the OS, so renaming "Table.jsx" to "table.jsx" doesnt do anything. This was causing some weird issues because I was updating the filenames and imports and had it working on my machine but git didnt pick up the changes. Ive just renamed the files again to be more descriptive